+++ CHANGES.md
@@ -5,6 +5,8 @@
   provides wrappers and conversion functions to and from time types.
 - Added a convenient `granularityToDiffTime` to convert from `Granularity` to
   `NominalDiffTime`.
+- Fixed not being able to use a start/end time in conjunction with a count for
+  the candlestick endpoints.
 
 # 0.1.0.0